feat: Implement WnVoIoBM and WnVoIoBMFactory classes for object management
- Added WnVoIoBM class extending AbstractIoBM with methods for handle creation, object copying, and removal.
- Introduced WnVoIoBMFactory for loading WnIoBM instances with dependency injection for WnIo and WnIoHandleManager.
- Created AbstractIoVfsIndexer class to provide base functionality for VFS indexing.
- Developed WnVofsObj class implementing WnObj for virtual file system object representation.
- Added command documentation for @raw_msg filter in mailx.
- Implemented WnSimpleClientGetter for simplified client retrieval.
- Created WnWeirdTest suite to run various indexer tests.
- Added WnVofsIndexerTest for testing VofsIndexer functionality.

+++ src/com/site0/walnut/api/io/WnIo.java
@@ -134,7 +134,7 @@ public interface WnIo {
 
     List<WnObj> getChildren(WnObj o, String name);
 
-    long count(WnQuery q);
+    int count(WnQuery q);
 
     /**
      * 根据一个指定条件，对一个键进行聚集汇总计算
