+++ src/com/site0/walnut/core/io/AbstractWnIoWrapper.java
@@ -410,7 +410,7 @@ public abstract class AbstractWnIoWrapper implements WnIo {
         return io.getChildren(o, name);
     }
 
-    public long count(WnQuery q) {
+    public int count(WnQuery q) {
         return io.count(q);
     }
